update profile screen begun - Edit email and re-authenticate Google mail flow has been added.
